5.gün

+++ northwind/src/main/java/com/kodlama/io/northwind/entities/Product.java
@@ -5,7 +5,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
@@ -31,7 +30,7 @@ public class Product {
 	private double unitPrice;
 	
 	@Column(name="unitsInStock")
-	private double unitsInStock;
+	private int unitsInStock;
 	
 	@ManyToOne
 //	@JoinColumn(name="category_id")
